+++ src/newcharacter.cpp
@@ -2419,7 +2419,7 @@ tab_direction set_description( const catacurses::window &w, avatar &you, const b
             redraw = true;
         } else if( action == "ANY_INPUT" &&
                    !MAP_SHARING::isSharing() ) { // Don't edit names when sharing maps
-            const long ch = ctxt.get_raw_input().get_first_input();
+            const int ch = ctxt.get_raw_input().get_first_input();
             utf8_wrapper wrap( you.name );
             if( ch == KEY_BACKSPACE ) {
                 if( !wrap.empty() ) {
@@ -2519,7 +2519,7 @@ trait_id Character::random_bad_trait()
 
 cata::optional<std::string> query_for_template_name()
 {
-    static const std::set<long> fname_char_blacklist = {
+    static const std::set<int> fname_char_blacklist = {
 #if defined(_WIN32)
         '\"', '*', '/', ':', '<', '>', '?', '\\', '|',
         '\x01', '\x02', '\x03', '\x04', '\x05', '\x06', '\x07',         '\x09',
@@ -2537,7 +2537,7 @@ cata::optional<std::string> query_for_template_name()
     spop.title( title );
     spop.description( desc );
     spop.width( FULL_SCREEN_WIDTH - utf8_width( title ) - 8 );
-    for( long character : fname_char_blacklist ) {
+    for( int character : fname_char_blacklist ) {
         spop.callbacks[ character ] = []() {
             return true;
         };
